+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Remove unused `REACT_APP_LOGIN_ENABLED` env var (#1006)
+- Fix infinite remix loop when `BYPASS_AUTH` set in `editor-api` (#1007)
 
 ## [0.23.0] - 2024-05-09